backend/pyproject.toml
[project]
name = "appartment-agent-backend"
version = "2.0.0"
description = "AI-powered apartment purchasing decision platform with LangGraph agent"
requires-python = "==3.10.*"
license = {text = "MIT"}
authors = [
    {name = "Appartment Agent Team"}
]

dependencies = [
    # FastAPI & Web Framework
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.2",
    # Database
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    # Data & Validation
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pandas>=2.1.3",
    "numpy>=1.26.2",
    "openpyxl>=3.1.2",
    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1,<4.1.0", # Pin to 4.0.x for passlib compatibility
    # AI & LLM (LangGraph Stack)
    "anthropic>=0.75.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.2.0",
    # Document Processing
    "pypdf2>=3.0.1",
    "pymupdf>=1.23.0",
    "pillow>=10.1.0",
    # Workflow & Storage (Temporal + MinIO)
    "temporalio>=1.5.0",
    "minio>=7.2.0",
    "redis>=5.0.1",
    "celery>=5.3.4",
    # Utilities
    "python-dotenv>=1.0.0",
    "grandalf>=0.8",
    "google-genai>=1.0.0",
    "logfire>=4.19.0",
    "langchain-google-genai>=4.2.0",
    # Google Cloud (for GCP deployment)
    "google-cloud-storage>=2.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.287",
    "mypy>=1.5.0",
]

viz = [
    # Graph visualization (optional - requires system graphviz)
    "pygraphviz>=1.11",
    "graphviz>=0.20.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.287",
]

# UV configuration
# Note: UV uses PyPI by default, no need to specify index

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
